move vizualization scripts to analysis
